Split test case into two test cases
