+++ src/python_testing/TC_MCORE_FS_1_3.py
@@ -27,7 +27,7 @@
 # test-runner-run/run1/app: examples/fabric-admin/scripts/fabric-sync-app.py
 # test-runner-run/run1/app-args: --app-admin=${FABRIC_ADMIN_APP} --app-bridge=${FABRIC_BRIDGE_APP} --stdin-pipe=dut-fsa/stdin --storage-dir=dut-fsa --discriminator=1234
 # test-runner-run/run1/factoryreset: True
-# test-runner-run/run1/script-args: --storage-path admin_storage.json --commissioning-method on-network --discriminator 1234 --passcode 20202021 --string-arg th_fsa_admin_path:${FABRIC_ADMIN_APP} th_fsa_bridge_path:${FABRIC_BRIDGE_APP} th_server_app_path:${LIGHTING_APP_NO_UNIQUE_ID} dut_fsa_stdin_pipe:dut-fsa/stdin
+# test-runner-run/run1/script-args: --storage-path admin_storage.json --commissioning-method on-network --discriminator 1234 --passcode 20202021 --string-arg th_server_app_path:${LIGHTING_APP_NO_UNIQUE_ID}
 # test-runner-run/run1/script-start-delay: 10
 # test-runner-run/run1/quiet: false
 # === END CI TEST ARGUMENTS ===
@@ -99,78 +99,16 @@ class ThreadWithStop(threading.Thread):
         self.join()
 
 
-class FabricSyncApp:
-
-    APP_PATH = "examples/fabric-admin/scripts/fabric-sync-app.py"
-    FABRIC_ADMIN_PATH = "out/linux-x64-fabric-admin-rpc/fabric-admin"
-    FABRIC_BRIDGE_PATH = "out/linux-x64-fabric-bridge-rpc/fabric-bridge-app"
-
-    def _process_admin_output(self, line):
-        if self.wait_for_text_text is not None and self.wait_for_text_text in line:
-            self.wait_for_text_event.set()
-
-    def wait_for_text(self, timeout=30):
-        if not self.wait_for_text_event.wait(timeout=timeout):
-            raise Exception(f"Timeout waiting for text: {self.wait_for_text_text}")
-        self.wait_for_text_event.clear()
-        self.wait_for_text_text = None
-
-    def __init__(self, storageDir, fabricName=None, nodeId=None, vendorId=None, paaTrustStorePath=None,
-                 bridgePort=None, bridgeDiscriminator=None, bridgePasscode=None):
-
-        self.wait_for_text_event = threading.Event()
-        self.wait_for_text_text = None
-
-        args = [self.APP_PATH]
-        args.append(f"--app-admin={self.FABRIC_ADMIN_PATH}")
-        args.append(f"--app-bridge={self.FABRIC_BRIDGE_PATH}")
-        # Override default ports, so it will be possible to run
-        # our TH_FSA alongside the DUT_FSA during CI testing.
-        args.append("--app-admin-rpc-port=44000")
-        args.append("--app-bridge-rpc-port=44001")
-        # Keep the storage directory in a temporary location.
-        args.append(f"--storage-dir={storageDir}")
-        # FIXME: Passing custom PAA store breaks something
-        # if paaTrustStorePath is not None:
-        #     args.append(f"--paa-trust-store-path={paaTrustStorePath}")
-        if fabricName is not None:
-            args.append(f"--commissioner-name={fabricName}")
-        if nodeId is not None:
-            args.append(f"--commissioner-node-id={nodeId}")
-        args.append(f"--commissioner-vendor-id={vendorId}")
-        args.append(f"--secured-device-port={bridgePort}")
-        args.append(f"--discriminator={bridgeDiscriminator}")
-        args.append(f"--passcode={bridgePasscode}")
-
-        self.admin = ThreadWithStop(args, stdout_cb=self._process_admin_output)
-        self.wait_for_text_text = "Connected to Fabric-Bridge"
-        self.admin.start()
-
-        # Wait for the bridge to connect to the admin.
-        self.wait_for_text()
-
-    def CommissionOnNetwork(self, nodeId, setupPinCode=None, filterType=None, filter=None):
-        self.wait_for_text_text = f"Commissioning complete for node ID 0x{nodeId:016x}: success"
-        # Send the commissioning command to the admin.
-        self.admin.p.stdin.write(f"pairing onnetwork {nodeId} {setupPinCode}\n")
-        self.admin.p.stdin.flush()
-        # Wait for success message.
-        self.wait_for_text()
-
-    def stop(self):
-        self.admin.stop()
-
-
 class AppServer:
 
-    def __init__(self, tag, app, storageDir, port=None, discriminator=None, passcode=None):
+    def __init__(self, app, storageDir, port=None, discriminator=None, passcode=None):
 
         args = [app]
         args.extend(["--KVS", tempfile.mkstemp(dir=storageDir, prefix="kvs-app-")[1]])
         args.extend(['--secured-device-port', str(port)])
         args.extend(["--discriminator", str(discriminator)])
         args.extend(["--passcode", str(passcode)])
-        self.app = ThreadWithStop(args, tag=tag)
+        self.app = ThreadWithStop(args, tag="APP")
         self.app.start()
 
     def stop(self):
@@ -187,18 +125,6 @@ class TC_MCORE_FS_1_3(MatterBaseTest):
     def setup_class(self):
         super().setup_class()
 
-        # Get the path to the TH_FSA (fabric-admin and fabric-bridge) app from
-        # the user params or use the default path.
-        FabricSyncApp.APP_PATH = self.user_params.get("th_fsa_app_path", FabricSyncApp.APP_PATH)
-        if not os.path.exists(FabricSyncApp.APP_PATH):
-            asserts.fail("This test requires a TH_FSA app. Specify app path with --string-arg th_fsa_app_path:<path_to_app>")
-        FabricSyncApp.FABRIC_ADMIN_PATH = self.user_params.get("th_fsa_admin_path", FabricSyncApp.FABRIC_ADMIN_PATH)
-        if not os.path.exists(FabricSyncApp.FABRIC_ADMIN_PATH):
-            asserts.fail("This test requires a TH_FSA_ADMIN app. Specify app path with --string-arg th_fsa_admin_path:<path_to_app>")
-        FabricSyncApp.FABRIC_BRIDGE_PATH = self.user_params.get("th_fsa_bridge_path", FabricSyncApp.FABRIC_BRIDGE_PATH)
-        if not os.path.exists(FabricSyncApp.FABRIC_BRIDGE_PATH):
-            asserts.fail("This test requires a TH_FSA_BRIDGE app. Specify app path with --string-arg th_fsa_bridge_path:<path_to_app>")
-
         # Get the path to the TH_SERVER app from the user params.
         th_server_app = self.user_params.get("th_server_app_path", None)
         if not th_server_app:
@@ -210,54 +136,20 @@ class TC_MCORE_FS_1_3(MatterBaseTest):
         self.storage = tempfile.TemporaryDirectory(prefix=self.__class__.__name__)
         logging.info("Temporary storage directory: %s", self.storage.name)
 
-        self.th_fsa_bridge_address = "::1"
-        self.th_fsa_bridge_port = 5543
-        self.th_fsa_bridge_discriminator = random.randint(0, 4095)
-        self.th_fsa_bridge_passcode = 20202021
+        self.th_server_port = 5544
+        self.th_server_discriminator = random.randint(0, 4095)
+        self.th_server_passcode = 20202021
 
-        self.th_fsa_controller = FabricSyncApp(
-            storageDir=self.storage.name,
-            paaTrustStorePath=self.matter_test_config.paa_trust_store_path,
-            bridgePort=self.th_fsa_bridge_port,
-            bridgeDiscriminator=self.th_fsa_bridge_discriminator,
-            bridgePasscode=self.th_fsa_bridge_passcode,
-            vendorId=0xFFF1)
-
-        # Get the named pipe path for the DUT_FSA app input from the user params.
-        dut_fsa_stdin_pipe = self.user_params.get("dut_fsa_stdin_pipe", None)
-        if dut_fsa_stdin_pipe is not None:
-            self.dut_fsa_stdin = open(dut_fsa_stdin_pipe, "w")
-
-        self.server_for_dut_port = 5544
-        self.server_for_dut_discriminator = random.randint(0, 4095)
-        self.server_for_dut_passcode = 20202021
-
-        # Start the TH_SERVER_FOR_DUT_FSA app.
-        self.server_for_dut = AppServer(
-            "APP1",
+        # Start the TH_SERVER app.
+        self.th_server = AppServer(
             th_server_app,
             storageDir=self.storage.name,
-            port=self.server_for_dut_port,
-            discriminator=self.server_for_dut_discriminator,
-            passcode=self.server_for_dut_passcode)
-
-        self.server_for_th_port = 5545
-        self.server_for_th_discriminator = random.randint(0, 4095)
-        self.server_for_th_passcode = 20202021
-
-        # Start the TH_SERVER_FOR_TH_FSA app.
-        self.server_for_th = AppServer(
-            "APP2",
-            th_server_app,
-            storageDir=self.storage.name,
-            port=self.server_for_th_port,
-            discriminator=self.server_for_th_discriminator,
-            passcode=self.server_for_th_passcode)
+            port=self.th_server_port,
+            discriminator=self.th_server_discriminator,
+            passcode=self.th_server_passcode)
 
     def teardown_class(self):
-        self.th_fsa_controller.stop()
-        self.server_for_dut.stop()
-        self.server_for_th.stop()
+        self.th_server.stop()
         self.storage.cleanup()
         super().teardown_class()
 
@@ -330,30 +222,29 @@ class TC_MCORE_FS_1_3(MatterBaseTest):
     @async_test_body
     async def test_TC_MCORE_FS_1_3(self):
         self.is_ci = self.check_pics('PICS_SDK_CI_ONLY')
-        self.is_ci = True
 
         # Commissioning - done
         self.step(0)
 
-        th_server_for_dut_fsa_th_node_id = 1
-        self.print_step(1, "Commissioning TH_SERVER_FOR_DUT_FSA to TH fabric")
+        th_server_th_node_id = 1
+        self.print_step(1, "Commissioning TH_SERVER to TH fabric")
         await self.default_controller.CommissionOnNetwork(
-            nodeId=th_server_for_dut_fsa_th_node_id,
-            setupPinCode=self.server_for_dut_passcode,
+            nodeId=th_server_th_node_id,
+            setupPinCode=self.th_server_passcode,
             filterType=ChipDeviceCtrl.DiscoveryFilterType.LONG_DISCRIMINATOR,
-            filter=self.server_for_dut_discriminator,
+            filter=self.th_server_discriminator,
         )
 
-        self.print_step(2, "Verify that TH_SERVER_FOR_DUT_FSA does not have a UniqueID")
+        self.print_step(2, "Verify that TH_SERVER does not have a UniqueID")
         # FIXME: Sometimes reading the UniqueID fails...
         await self.read_single_attribute_expect_error(
             cluster=Clusters.BasicInformation,
             attribute=Clusters.BasicInformation.Attributes.UniqueID,
-            node_id=th_server_for_dut_fsa_th_node_id,
+            node_id=th_server_th_node_id,
             error=Status.UnsupportedAttribute,
         )
 
-        # Get the list of endpoints on the DUT_FSA_BRIDGE before adding the TH_SERVER_FOR_DUT_FSA.
+        # Get the list of endpoints on the DUT_FSA_BRIDGE before adding the TH_SERVER.
         dut_fsa_bridge_endpoints = set(await self.read_single_attribute_check_success(
             cluster=Clusters.Descriptor,
             attribute=Clusters.Descriptor.Attributes.PartsList,
@@ -361,15 +252,15 @@ class TC_MCORE_FS_1_3(MatterBaseTest):
             endpoint=0,
         ))
 
-        self.print_step(3, "Commissioning TH_SERVER_FOR_DUT_FSA to DUT_FSA fabric")
+        self.print_step(3, "Commissioning TH_SERVER to DUT_FSA fabric")
         await self.commission_via_commissioner_control(
             controller_node_id=self.dut_node_id,
-            device_node_id=th_server_for_dut_fsa_th_node_id)
+            device_node_id=th_server_th_node_id)
 
         # Wait for the device to appear on the DUT_FSA_BRIDGE.
         await asyncio.sleep(2)
 
-        # Get the list of endpoints on the DUT_FSA_BRIDGE after adding the TH_SERVER_FOR_DUT_FSA.
+        # Get the list of endpoints on the DUT_FSA_BRIDGE after adding the TH_SERVER.
         dut_fsa_bridge_endpoints_new = set(await self.read_single_attribute_check_success(
             cluster=Clusters.Descriptor,
             attribute=Clusters.Descriptor.Attributes.PartsList,
@@ -377,153 +268,22 @@ class TC_MCORE_FS_1_3(MatterBaseTest):
             endpoint=0,
         ))
 
-        # Get the endpoint number for just added TH_SERVER_FOR_DUT_FSA.
-        logging.info("Endpoints on DUT_FSA_BRIDGE: old=%s, new=%s", dut_fsa_bridge_endpoints, dut_fsa_bridge_endpoints_new)
-        asserts.assert_true(dut_fsa_bridge_endpoints_new.issuperset(dut_fsa_bridge_endpoints),
-                            "Expected only new endpoints to be added")
-        unique_endpoints_set = dut_fsa_bridge_endpoints_new - dut_fsa_bridge_endpoints
-        asserts.assert_equal(len(unique_endpoints_set), 1, "Expected only one new endpoint on DUT_FSA")
-        dut_fsa_bridge_dut_server_endpoint = list(unique_endpoints_set)[0]
-        dut_fsa_bridge_endpoints = dut_fsa_bridge_endpoints_new
-
-        self.print_step(4, "Verify that DUT_FSA created a UniqueID for TH_SERVER_FOR_DUT_FSA")
-        dut_fsa_bridge_dut_server_unique_id = await self.read_single_attribute_check_success(
-            cluster=Clusters.BridgedDeviceBasicInformation,
-            attribute=Clusters.BridgedDeviceBasicInformation.Attributes.UniqueID,
-            endpoint=dut_fsa_bridge_dut_server_endpoint)
-        asserts.assert_true(type_matches(dut_fsa_bridge_dut_server_unique_id, str), "UniqueID should be a string")
-        asserts.assert_true(dut_fsa_bridge_dut_server_unique_id, "UniqueID should not be an empty string")
-        logging.info("TH_SERVER_FOR_DUT_FSA on DUT_SERVER_BRIDGE UniqueID: %s", dut_fsa_bridge_dut_server_unique_id)
-
-        th_fsa_bridge_th_node_id = 2
-        self.print_step(5, "Commissioning TH_FSA_BRIDGE to TH fabric")
-        await self.default_controller.CommissionOnNetwork(
-            nodeId=th_fsa_bridge_th_node_id,
-            setupPinCode=self.th_fsa_bridge_passcode,
-            filterType=ChipDeviceCtrl.DiscoveryFilterType.LONG_DISCRIMINATOR,
-            filter=self.th_fsa_bridge_discriminator,
-        )
-
-        # Get the list of endpoints on the TH_FSA_BRIDGE before adding the TH_SERVER_FOR_TH_FSA.
-        th_fsa_bridge_endpoints = set(await self.read_single_attribute_check_success(
-            cluster=Clusters.Descriptor,
-            attribute=Clusters.Descriptor.Attributes.PartsList,
-            node_id=th_fsa_bridge_th_node_id,
-            endpoint=0,
-        ))
-
-        th_server_for_th_fsa_th_fsa_node_id = 3
-        self.print_step(6, "Commissioning TH_SERVER_FOR_TH_FSA to TH_FSA")
-        self.th_fsa_controller.CommissionOnNetwork(
-            nodeId=th_server_for_th_fsa_th_fsa_node_id,
-            setupPinCode=self.server_for_th_passcode,
-            filterType=ChipDeviceCtrl.DiscoveryFilterType.LONG_DISCRIMINATOR,
-            filter=self.server_for_th_discriminator,
-        )
-
-        # Wait some time, so the dynamic endpoint will appear on the TH_FSA_BRIDGE.
-        await asyncio.sleep(5)
-
-        # Get the list of endpoints on the TH_FSA_BRIDGE after adding the TH_SERVER_FOR_TH_FSA.
-        th_fsa_bridge_endpoints_new = set(await self.read_single_attribute_check_success(
-            cluster=Clusters.Descriptor,
-            attribute=Clusters.Descriptor.Attributes.PartsList,
-            node_id=th_fsa_bridge_th_node_id,
-            endpoint=0,
-        ))
-
-        # Get the endpoint number for just added TH_SERVER_FOR_TH_FSA.
-        logging.info("Endpoints on TH_FSA_BRIDGE: old=%s, new=%s", th_fsa_bridge_endpoints, th_fsa_bridge_endpoints_new)
-        asserts.assert_true(th_fsa_bridge_endpoints_new.issuperset(th_fsa_bridge_endpoints),
-                            "Expected only new endpoints to be added")
-        unique_endpoints_set = th_fsa_bridge_endpoints_new - th_fsa_bridge_endpoints
-        asserts.assert_equal(len(unique_endpoints_set), 1, "Expected only one new endpoint")
-        th_fsa_bridge_th_server_endpoint = list(unique_endpoints_set)[0]
-
-        self.print_step(7, "Verify that TH_FSA created a UniqueID for TH_SERVER_FOR_TH_FSA")
-        th_fsa_bridge_th_server_unique_id = await self.read_single_attribute_check_success(
-            cluster=Clusters.BridgedDeviceBasicInformation,
-            attribute=Clusters.BridgedDeviceBasicInformation.Attributes.UniqueID,
-            node_id=th_fsa_bridge_th_node_id,
-            endpoint=th_fsa_bridge_th_server_endpoint)
-        asserts.assert_true(type_matches(th_fsa_bridge_th_server_unique_id, str), "UniqueID should be a string")
-        asserts.assert_true(th_fsa_bridge_th_server_unique_id, "UniqueID should not be an empty string")
-        logging.info("TH_SERVER_FOR_TH_FSA on TH_SERVER_BRIDGE UniqueID: %s", th_fsa_bridge_th_server_unique_id)
-
-        # Make sure that the UniqueID on the TH_FSA_BRIDGE is different from the one on the DUT_FSA_BRIDGE.
-        asserts.assert_not_equal(dut_fsa_bridge_dut_server_unique_id, th_fsa_bridge_th_server_unique_id,
-                                 "UniqueID on DUT_FSA_BRIDGE and TH_FSA_BRIDGE should be different")
-
-        discriminator = random.randint(0, 4095)
-        self.print_step(8, "Open commissioning window on TH_FSA_BRIDGE")
-        params = await self.default_controller.OpenCommissioningWindow(
-            nodeid=th_fsa_bridge_th_node_id,
-            option=self.default_controller.CommissioningWindowPasscode.kTokenWithRandomPin,
-            discriminator=discriminator,
-            iteration=10000,
-            timeout=600)
-
-        self.print_step(9, "Commissioning TH_FSA_BRIDGE to DUT_FSA fabric")
-        if not self.is_ci:
-            self.wait_for_user_input(
-                f"Commission TH Fabric-Sync Bridge on DUT using manufacturer specified mechanism.\n"
-                f"Use the following parameters:\n"
-                f"- discriminator: {discriminator}\n"
-                f"- setupPinCode: {params.setupPinCode}\n"
-                f"- setupQRCode: {params.setupQRCode}\n"
-                f"- setupManualCode: {params.setupManualCode}\n"
-                f"If using FabricSync Admin, you may type:\n"
-                f">>> fabricsync add-bridge <desired_node_id> {params.setupPinCode} <th_host_ip> {self.th_fsa_bridge_port}")
-        else:
-            self.dut_fsa_stdin.write(
-                f"fabricsync add-bridge 10 {params.setupPinCode} {self.th_fsa_bridge_address} {self.th_fsa_bridge_port}\n")
-            self.dut_fsa_stdin.flush()
-            # Wait for the commissioning to complete.
-            await asyncio.sleep(5)
-
-        self.print_step(10, "Synchronize TH_SERVER_FOR_TH_FSA from TH_FSA to DUT_FSA fabric")
-        if not self.is_ci:
-            self.wait_for_user_input(
-                f"Synchronize endpoint from TH Fabric-Sync Bridge to DUT using manufacturer specified mechanism.\n"
-                f"Use the following parameters:\n"
-                f"- bridgeDynamicEndpoint: {th_fsa_bridge_th_server_endpoint}\n"
-                f"If using FabricSync Admin, you may type:\n"
-                f">>> fabricsync sync-device {th_fsa_bridge_th_server_endpoint}")
-        else:
-            self.dut_fsa_stdin.write(f"fabricsync sync-device {th_fsa_bridge_th_server_endpoint}\n")
-            self.dut_fsa_stdin.flush()
-            # Wait for the synchronization to complete.
-            await asyncio.sleep(5)
-
-        # Get the list of endpoints on the DUT_FSA_BRIDGE after synchronization
-        dut_fsa_bridge_endpoints_new = set(await self.read_single_attribute_check_success(
-            cluster=Clusters.Descriptor,
-            attribute=Clusters.Descriptor.Attributes.PartsList,
-            node_id=self.dut_node_id,
-            endpoint=0,
-        ))
-
-        # Get the endpoint number for just synced TH_SERVER_FOR_TH_FSA.
+        # Get the endpoint number for just added TH_SERVER.
         logging.info("Endpoints on DUT_FSA_BRIDGE: old=%s, new=%s", dut_fsa_bridge_endpoints, dut_fsa_bridge_endpoints_new)
         asserts.assert_true(dut_fsa_bridge_endpoints_new.issuperset(dut_fsa_bridge_endpoints),
                             "Expected only new endpoints to be added")
         unique_endpoints_set = dut_fsa_bridge_endpoints_new - dut_fsa_bridge_endpoints
         asserts.assert_equal(len(unique_endpoints_set), 1, "Expected only one new endpoint on DUT_FSA")
         dut_fsa_bridge_th_server_endpoint = list(unique_endpoints_set)[0]
-        dut_fsa_bridge_endpoints = dut_fsa_bridge_endpoints_new
 
-        self.print_step(11, "Verify that DUT_FSA copied the TH_SERVER_FOR_TH_FSA UniqueID from TH_FSA")
+        self.print_step(4, "Verify that DUT_FSA created a UniqueID for TH_SERVER")
         dut_fsa_bridge_th_server_unique_id = await self.read_single_attribute_check_success(
             cluster=Clusters.BridgedDeviceBasicInformation,
             attribute=Clusters.BridgedDeviceBasicInformation.Attributes.UniqueID,
             endpoint=dut_fsa_bridge_th_server_endpoint)
         asserts.assert_true(type_matches(dut_fsa_bridge_th_server_unique_id, str), "UniqueID should be a string")
         asserts.assert_true(dut_fsa_bridge_th_server_unique_id, "UniqueID should not be an empty string")
-        logging.info("TH_SERVER_FOR_TH_FSA on DUT_SERVER_BRIDGE UniqueID: %s", dut_fsa_bridge_th_server_unique_id)
-
-        # Make sure that the UniqueID on the DUT_FSA_BRIDGE is the same as the one on the DUT_FSA_BRIDGE.
-        asserts.assert_equal(dut_fsa_bridge_th_server_unique_id, th_fsa_bridge_th_server_unique_id,
-                             "UniqueID on DUT_FSA_BRIDGE and TH_FSA_BRIDGE should be the same")
+        logging.info("UniqueID generated for TH_SERVER: %s", dut_fsa_bridge_th_server_unique_id)
 
 
 if __name__ == "__main__":
